refactor(st-09057): modularize transaction flow
