+++ planning/checklists/epic-09-story-tasks.md
@@ -2288,23 +2288,41 @@ Implementation notes:
 **Branch:** `refactor/st-09057-transaction-flow-modularization`
 
 ### Checklist
-- [ ] Create branch `refactor/st-09057-transaction-flow-modularization`
+- [x] Create branch `refactor/st-09057-transaction-flow-modularization`
 - [ ] Create draft PR with story ID in title
-- [ ] Define test strategy before implementation: cover runtime modularization and test-file modularization; first failing test should prove transaction behavior remains stable while the oversized runtime and test files are split
-- [ ] Write or update the failing automated test before production changes when practical; if not practical, record why before implementation
-- [ ] Reduce `packages/tools/src/data/relational/query/transaction.ts` below the 300 line planning cutoff by extracting focused internal modules for transaction lifecycle, savepoint handling, timeout/rollback flow, and vendor/session helpers behind a stable facade
-- [ ] Keep extracted production modules below the 300 line planning cutoff as well; do not satisfy the story by only shrinking the public facade and moving the bulk into a new oversized helper unless an explicit exception is documented in the story notes
-- [ ] Split transaction coverage into focused test modules so lifecycle, timeout, and savepoint-path behavior no longer depend on a single oversized test surface
-- [ ] Preserve existing transaction behavior, rollback semantics, timeout handling, and public imports
-- [ ] Add/update production code until focused tests pass, keeping test evidence in checklist notes and PR body
-- [ ] Record explicit-`any` warning deltas and file-size/responsibility improvements for touched transaction modules in story docs
-- [ ] Add or update story documentation at `docs/st09057-transaction-flow-modularization.md` (or document why not required)
-- [ ] Assess residual test impact; add/update additional automated tests when needed, or document why no further tests are required
-- [ ] Run full test suite before finalizing the PR and record results
-- [ ] Run lint (`pnpm lint`) before finalizing the PR and record results
+- [x] Define test strategy before implementation: cover runtime modularization and test-file modularization; first failing test should prove transaction behavior remains stable while the oversized runtime and test files are split
+- [x] Write or update the failing automated test before production changes when practical; if not practical, record why before implementation
+- [x] Reduce `packages/tools/src/data/relational/query/transaction.ts` below the 300 line planning cutoff by extracting focused internal modules for transaction lifecycle, savepoint handling, timeout/rollback flow, and vendor/session helpers behind a stable facade
+- [x] Keep extracted production modules below the 300 line planning cutoff as well; do not satisfy the story by only shrinking the public facade and moving the bulk into a new oversized helper unless an explicit exception is documented in the story notes
+- [x] Split transaction coverage into focused test modules so lifecycle, timeout, and savepoint-path behavior no longer depend on a single oversized test surface
+- [x] Preserve existing transaction behavior, rollback semantics, timeout handling, and public imports
+- [x] Add/update production code until focused tests pass, keeping test evidence in checklist notes and PR body
+- [x] Record explicit-`any` warning deltas and file-size/responsibility improvements for touched transaction modules in story docs
+- [x] Add or update story documentation at `docs/st09057-transaction-flow-modularization.md` (or document why not required)
+- [x] Assess residual test impact; add/update additional automated tests when needed, or document why no further tests are required
+- [x] Run full test suite before finalizing the PR and record results
+- [x] Run lint (`pnpm lint`) before finalizing the PR and record results
 - [ ] Commit completed checklist items as logical commits and push updates
 - [ ] Mark PR Ready only after all story tasks are complete
 - [ ] Wait for merge; do not merge directly from local branch
+
+Implementation notes:
+
+- Test-first rationale:
+  - A structure-only failing test would mostly prove file layout rather than transaction behavior. The practical test-first substitute was splitting the monolithic query transaction suite into focused lifecycle, option-handling, and savepoint suites first, then using those suites together with the existing relational integration coverage as the regression net during runtime extraction.
+- Focused package validation:
+  - `pnpm test --run packages/tools/tests/data/relational/query/transaction-lifecycle.test.ts packages/tools/tests/data/relational/query/transaction-options.test.ts packages/tools/tests/data/relational/query/transaction-savepoints.test.ts packages/tools/tests/data/relational/transaction.test.ts packages/tools/tests/data/relational/transaction-timeout-and-savepoint.test.ts` -> `4` files passed, `1` file skipped; `30` tests passed, `4` skipped
+  - `pnpm --filter @agentforge/tools typecheck`
+  - `pnpm --filter @agentforge/tools exec eslint src/data/relational/query/transaction.ts src/data/relational/query/transaction-types.ts src/data/relational/query/transaction-options.ts src/data/relational/query/transaction-managed.ts src/data/relational/query/transaction-runner.ts tests/data/relational/query/transaction-lifecycle.test.ts tests/data/relational/query/transaction-options.test.ts tests/data/relational/query/transaction-savepoints.test.ts tests/data/relational/query/transaction.test-utils.ts`
+- Baseline/quality gates:
+  - `pnpm lint:explicit-any:baseline` -> `84/289` workspace, `tools 53/67`
+- Full workspace validation:
+  - `pnpm test --run` -> `207` files passed, `18` skipped; `2307` tests passed, `286` skipped
+  - `pnpm lint` -> exit `0`; warnings only (`0` errors)
+- Residual test impact:
+  - No additional automated tests were needed beyond the focused query transaction split because lifecycle, option handling, timeout cancellation, and savepoint behavior are already covered by the new focused unit suites plus the existing relational integration tests. The full-suite count moved from `2313` to `2307` because the old monolithic query transaction test file was replaced by three narrower suites with less duplicated coverage while preserving the behavior contract.
+- CI impact assessment:
+  - No CI workflow change required. Existing package typecheck, focused tests, workspace test suite, lint, and explicit-`any` baseline gates already cover the extracted runtime and test surfaces.
 
 ---
